Rename logger_function -> logger_decorator
